+++ CHANGELOG.md
@@ -590,6 +590,8 @@
 * Added /randommusic, which plays a randomly chosen track from the player's current music list
 * Added /exit, which lets you exit the server from OOC
 * Server initiated messages will now attempt to include desks wherever possible
+* Added /zone_handicap and /zone_unhandicap. Players who enter an area part of a zone with a handicap will be subject to the imposed movement handicap automatically
+    - Also added /zone_handicap_add, which makes a player be subject to a zone handicap if their handicap was removed
 * Fixed scream_range in area list yaml files not supporting the keyword <ALL> to indicate all areas should be able to receive a scream coming from a particular area
 * Fixed scream_range in area list yaml files not checking if the areas a scream can reach to from a particular area exist
 * Fixed /scream, /whisper and /party_whisper raising errors if a message was sent to a deafened player with a bypass message starter. They now sent messages but filtered